requirements.txt
pandasai
streamlit
langchain_anthropic
seaborn
python-dotenv 
langchain_google_genai